Move SourceInfo and fix imports
Rename sourceinfo macro package
