+++ core/src/main/scala/chisel3/BlackBox.scala
@@ -7,7 +7,7 @@ import chisel3.internal.BaseBlackBox
 import chisel3.internal.Builder.pushCommand
 import chisel3.internal.firrtl._
 import chisel3.internal.throwException
-import chisel3.internal.sourceinfo.{SourceInfo, UnlocatableSourceInfo}
+import chisel3.experimental.{SourceInfo, UnlocatableSourceInfo}
 import scala.annotation.nowarn
 
 package internal {
