+++ core/src/main/scala/chisel3/Mem.scala
@@ -9,7 +9,8 @@ import firrtl.{ir => fir}
 import chisel3.internal._
 import chisel3.internal.Builder.pushCommand
 import chisel3.internal.firrtl._
-import chisel3.internal.sourceinfo.{MemTransform, SourceInfo, SourceInfoTransform, SourceLine, UnlocatableSourceInfo}
+import chisel3.internal.sourceinfomacros.{MemTransform, SourceInfoTransform}
+import chisel3.experimental.{SourceInfo, SourceLine, UnlocatableSourceInfo}
 
 object Mem {
