+++ core/src/main/scala/chisel3/Printf.scala
@@ -4,7 +4,7 @@ package chisel3
 
 import chisel3.internal._
 import chisel3.internal.Builder.pushCommand
-import chisel3.internal.sourceinfo.SourceInfo
+import chisel3.experimental.SourceInfo
 import scala.language.experimental.macros
 import scala.reflect.macros.blackbox
