+++ core/src/main/scala/chisel3/Reg.scala
@@ -7,7 +7,8 @@ import scala.language.experimental.macros
 import chisel3.internal._
 import chisel3.internal.Builder.pushCommand
 import chisel3.internal.firrtl._
-import chisel3.internal.sourceinfo.SourceInfo
+import chisel3.experimental.SourceInfo
+import chisel3.internal.sourceinfomacros.SourceInfoTransform
 
 /** Utility for constructing hardware registers
   *
