+++ core/src/main/scala/chisel3/SeqUtils.scala
@@ -2,11 +2,11 @@
 
 package chisel3
 
-import chisel3.experimental.FixedPoint
+import chisel3.experimental.{FixedPoint, SourceInfo}
 import chisel3.internal.{prefix, throwException}
 
 import scala.language.experimental.macros
-import chisel3.internal.sourceinfo._
+import chisel3.internal.sourceinfomacros._
 import chisel3.internal.plugin.autoNameRecursively
 
 private[chisel3] object SeqUtils {
