+++ core/src/main/scala/chisel3/connectable/Connection.scala
@@ -6,7 +6,7 @@ import chisel3.{Aggregate, BiConnectException, Data, DontCare, RawModule}
 import chisel3.internal.{prefix, BiConnect, Builder}
 import chisel3.internal.Builder.pushCommand
 import chisel3.internal.firrtl.DefInvalid
-import chisel3.internal.sourceinfo.{SourceInfo, UnlocatableSourceInfo}
+import chisel3.experimental.{SourceInfo, UnlocatableSourceInfo}
 import chisel3.experimental.{attach, Analog}
 import Alignment.matchingZipOfChildren
