+++ core/src/main/scala/chisel3/experimental/Analog.scala
@@ -3,7 +3,7 @@
 package chisel3.experimental
 
 import chisel3.internal.firrtl.Width
-import chisel3.internal.sourceinfo.SourceInfo
+import chisel3.experimental.SourceInfo
 import chisel3.internal._
 import chisel3.{
   ActualDirection,
