+++ core/src/main/scala/chisel3/experimental/Attach.scala
@@ -6,7 +6,7 @@ import chisel3.RawModule
 import chisel3.internal._
 import chisel3.internal.Builder.pushCommand
 import chisel3.internal.firrtl._
-import chisel3.internal.sourceinfo.SourceInfo
+import chisel3.experimental.SourceInfo
 
 object attach {
   // Exceptions that can be generated by attach
