+++ core/src/main/scala/chisel3/experimental/dataview/package.scala
@@ -4,7 +4,7 @@ package chisel3.experimental
 
 import chisel3._
 import chisel3.internal._
-import chisel3.internal.sourceinfo.SourceInfo
+import chisel3.experimental.SourceInfo
 
 import scala.annotation.{implicitNotFound, tailrec}
 import scala.collection.mutable
