+++ src/main/scala/chisel3/compatibility.scala
@@ -120,7 +120,7 @@ package object Chisel {
   @nowarn("msg=Chisel compatibility mode is deprecated")
   implicit class fromBitsable[T <: Data](data: T) {
     import chisel3.CompileOptions
-    import chisel3.internal.sourceinfo.SourceInfo
+    import chisel3.experimental.SourceInfo
 
     /** Creates an new instance of this type, unpacking the input Bits into
       * structured data.
@@ -143,7 +143,8 @@ package object Chisel {
   @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
   object Vec extends chisel3.VecFactory {
     import chisel3.CompileOptions
-    import chisel3.internal.sourceinfo._
+    import chisel3.experimental._
+    import chisel3.internal.sourceinfomacros.VecTransform
 
     @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
     @deprecated("Vec argument order should be size, t; this will be removed by the official release", "chisel3")
@@ -376,7 +377,7 @@ package object Chisel {
   type Mem[T <: Data] = chisel3.Mem[T]
 
   implicit class MemCompatibility(a: Mem.type) {
-    import chisel3.internal.sourceinfo.UnlocatableSourceInfo
+    import chisel3.experimental.UnlocatableSourceInfo
 
     @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
     def apply[T <: Data](t: T, size: BigInt)(implicit compileOptions: CompileOptions): Mem[T] =
@@ -394,7 +395,7 @@ package object Chisel {
   type SeqMem[T <: Data] = chisel3.SyncReadMem[T]
 
   implicit class SeqMemCompatibility(a: SeqMem.type) {
-    import chisel3.internal.sourceinfo.SourceInfo
+    import chisel3.experimental.SourceInfo
 
     @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
     def apply[T <: Data](
@@ -437,7 +438,7 @@ package object Chisel {
   @nowarn("msg=Chisel compatibility mode is deprecated")
   object Reg {
     import chisel3.CompileOptions
-    import chisel3.internal.sourceinfo.SourceInfo
+    import chisel3.experimental.SourceInfo
 
     // Passthrough for chisel3.Reg
     // Single-element constructor to avoid issues caused by null default args in a type
@@ -811,7 +812,7 @@ package object Chisel {
 
   @nowarn("msg=Chisel compatibility mode is deprecated")
   implicit class DataCompatibility(a: Data) {
-    import chisel3.internal.sourceinfo.DeprecatedSourceInfo
+    import chisel3.experimental.DeprecatedSourceInfo
 
     @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
     def toBits(implicit compileOptions: CompileOptions): UInt = a.do_asUInt(DeprecatedSourceInfo, compileOptions)
@@ -820,7 +821,7 @@ package object Chisel {
 
   @nowarn("msg=Chisel compatibility mode is deprecated")
   implicit class VecLikeCompatibility[T <: Data](a: VecLike[T]) {
-    import chisel3.internal.sourceinfo.DeprecatedSourceInfo
+    import chisel3.experimental.DeprecatedSourceInfo
 
     @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
     def read(idx: UInt)(implicit compileOptions: CompileOptions): T = a.do_apply(idx)(compileOptions)
@@ -833,7 +834,7 @@ package object Chisel {
 
   @nowarn("msg=Chisel compatibility mode is deprecated")
   implicit class BitsCompatibility(a: Bits) {
-    import chisel3.internal.sourceinfo.DeprecatedSourceInfo
+    import chisel3.experimental.DeprecatedSourceInfo
 
     @deprecated("Chisel compatibility mode is deprecated. Use the chisel3 package instead.", "Chisel 3.6")
     final def asBits(implicit compileOptions: CompileOptions): Bits = a.do_asUInt(DeprecatedSourceInfo, compileOptions)
