+++ src/main/scala/chisel3/util/Reg.scala
@@ -5,7 +5,8 @@ package chisel3.util
 import scala.language.experimental.macros
 
 import chisel3._
-import chisel3.internal.sourceinfo.{SourceInfo, SourceInfoTransform}
+import chisel3.experimental.SourceInfo
+import chisel3.internal.sourceinfomacros.SourceInfoTransform
 
 object RegEnable {
